requirements.txt
# Core ML (FIXED)
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26,<2.0
nltk==3.8.1
joblib==1.3.2

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# MLOps
mlflow==2.9.2
dvc==3.30.1

# Monitoring
prometheus-client==0.19.0

# Utilities
pyyaml==6.0.1
requests==2.31.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
httpx==0.25.2